fix:企业大屏处理用户权限

+++ src/main/java/cn/shmedo/monitor/monibotbaseapi/model/response/dashboard/ResourceOverviewRes.java
@@ -18,7 +18,7 @@ import lombok.experimental.Accessors;
 @NoArgsConstructor
 @AllArgsConstructor
 public class ResourceOverviewRes {
-    private int companyCount;
-    private int areaCount;
-    private int projectCount;
+    private long companyCount;
+    private long areaCount;
+    private long projectCount;
 }
